Search and Drafting

+++ django_apps/posts/views.py
@@ -1,11 +1,13 @@
 # -*- coding: utf-8 -*-
 from __future__ import unicode_literals
-from django.http import HttpResponse, HttpResponseRedirect
+from django.http import HttpResponse, HttpResponseRedirect, Http404
 from django.shortcuts import render, get_object_or_404, redirect
 from .models import Post
 from django.contrib import messages
 from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
 from django.shortcuts import render
+from django.utils import timezone
+from django.db.models import Q
 
 from rest_framework.generics import (
 	ListAPIView,
@@ -20,15 +22,19 @@ from .forms import PostForm
 # Create your views here.
 
 def post_create(request): #C
+	if not request.user.is_staff or not request.user.is_superuser:
+		raise Http404
 	form = PostForm(request.POST or None, request.FILES or None) #Built in validation from forms
 	if form.is_valid(): #if all fields filled out then save
 		# print(form.cleaned_data.get("title")) #print field
 		instance = form.save(commit=False)
+		instance.user = request.user
 		instance.save()
 		messages.success(request, "Sucessfully Created")
 		return HttpResponseRedirect(instance.get_absolute_url())
 	context = {
 		"form": form,
+		"type": "Create",
 	}
 	# form = PostForm()
 	# if request.method == "POST":
@@ -39,6 +45,13 @@ def post_create(request): #C
 
 def post_detail(request, slug=None): #R
 	instance = get_object_or_404(Post, slug=slug)
+	if instance.draft:
+		if not request.user.is_staff or not request.user.is_superuser:
+			raise Http404
+	try:
+		instance = Post.objects.get(slug=slug)
+	except Post.DoesNotExist:
+		raise Http404("no matching!!")
 	context = {
 		"title": instance.title,
 		"instance": instance,
@@ -46,10 +59,20 @@ def post_detail(request, slug=None): #R
 	return render(request, "post_detail.html", context)
 
 def post_list(request):
-	queryset_list = Post.objects.all()
-	paginator = Paginator(queryset_list, 7) # Show 25 contacts per page
-	page = request.GET.get('page')
-	page_request_var = 'page'
+	# queryset_list = Post.objects.filter(draft=False).filter(publish__lte=timezone.now())
+	today = timezone.now().date
+	queryset_list = Post.objects.active()
+	if request.user.is_staff or request.user.is_superuser: #use to control display of view Edit and Delete buttons
+		queryset_list = Post.objects.all()
+	query = request.GET.get("q")
+	if query:
+		queryset_list = queryset_list.filter(
+			Q(title__icontains=query)| #import Q to seach multiple
+			Q(content__icontains=query)
+		).distinct() #No duplicte items
+	paginator = Paginator(queryset_list, 4) # Show 7 contacts per page
+	page = request.GET.get('abc') #must be named same as page_var
+	page_request_var = 'abc'
 	try:
 		queryset = paginator.page(page)
 	except PageNotAnInteger:
@@ -60,10 +83,13 @@ def post_list(request):
 		"object_list": queryset,
 		"title":"List",
 		"page_request_var": page_request_var,
+		"today" : today,
 	}
 	return render(request, "post_list.html", context)
 
 def post_update(request, slug=None): #U
+	if not request.user.is_staff or not request.user.is_superuser:
+		raise Http404
 	instance = get_object_or_404(Post, slug=slug)
 	form = PostForm(request.POST or None, request.FILES or None, instance=instance)
 	if form.is_valid():
@@ -75,10 +101,13 @@ def post_update(request, slug=None): #U
 		"title": instance.title,
 		"instance": instance,
 		"form": form,
+		"type": "Update",
 	}
 	return render(request, "post_form.html", context)
 
 def post_delete(request, slug=None): #D
+	if not request.user.is_staff or not request.user.is_superuser:
+		raise Http404
 	instance = get_object_or_404(Post, slug=slug)
 	instance.delete()
 	messages.success(request, "Sucessfully Deleted", extra_tags='html_safe')
@@ -94,7 +123,7 @@ def post_home(request): #C
 class PostListAPIView(ListAPIView):
 	serializer_class = PostListSerializer #Turn model into Json
 	queryset = Post.objects.all() #must have queryset or overide get_querset() method
-	lookup_field = 'id'
+	lookup_field = 'slug'
 
 	# search_fields = ['title']
 	# def get_queryset(self, *args, **kwargs):
@@ -108,5 +137,5 @@ class PostListAPIView(ListAPIView):
 class PostDetailAPIView(RetrieveAPIView):
 	serializer_class = PostDetailSerializer
 	queryset = Post.objects.all()
-	lookup_field = 'id'
+	lookup_field = 'slug'
